src/main/java/org/example/leetcode/Linked_List_Cycle.java
package org.example.leetcode;

import java.util.HashSet;
import java.util.Set;

public class Linked_List_Cycle {

    /*
    Given head, the head of a linked list, determine if the linked list has a cycle in it.

    There is a cycle in a linked list if there is some node in the list that can be reached again by continuously following the next pointer.
    Internally, pos is used to denote the index of the node that tail's next pointer is connected to. Note that pos is not passed as a parameter.

    Return true if there is a cycle in the linked list. Otherwise, return false

    Constraints:
    The number of the nodes in the list is in the range [0, 104].
    -105 <= Node.val <= 105
    pos is -1 or a valid index in the linked-list.
     */
    public boolean hasCycle(ListNode head) {
        if (head == null) {
            return false;
        }
        Set<ListNode> set = new HashSet<>();
        set.add(head);
        ListNode tmp = head;
        while (tmp.next != null) {
            tmp = tmp.next;
            if (set.contains(tmp)) {
                return true;
            }
            set.add(tmp);
        }
        return false;
    }

    static class ListNode {
        int val;
        ListNode next;

        ListNode(int x) {
            val = x;
            next = null;
        }
    }
}
